Adding manual unlink to cleanup temp clipboard file.
mkstemp and fdopen(..., "wD") aren't removing the temporary file as
expected. I don't know if I'm doing something wrong or what, but the easy
solution (again, probably not entirely portable) is a manual call to
unlink. ;) We shall probably want to separate the file IO into separate
routines so it can be reused in the Get function...